requirements.txt
flask
opencv-python
mediapipe
numpy
pyserial
